implement optional trimming of whitespace from secrets
This adds a new configuration option called `trim` which, when set, trims
all leading and trailing whitespace from secrets.
This can be useful because some editors automatically write an extra newline at the end of
each file during editing which is often not desired when that files content is intended
as a password or other secret.

Closes https://github.com/lilioid/nm-file-secret-agent/issues/6